requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
openai
requests
requests-negotiate-sspi
truststore
python-dotenv
numpy
scikit-learn
openpyxl